day 5/Financial Intelligence Platform/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
langchain>=0.1.16
langgraph>=0.0.35
crewai==0.28.8
chromadb>=0.4.24
sentence-transformers>=2.6.1
python-dotenv>=1.0.0
groq>=0.5.0
langchain-groq>=0.1.2
langchain-openai>=0.1.3
python-multipart>=0.0.9
